# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Octacer Internal Tooling Platform",
    "description": "Octacer's internal tooling platform — a single API bundling attendance report generation, document & text extraction (OCR), image processing, audio/video transcription, and web automation.",
    "version": "1.0.0"
  },
  "paths": {
    "/generate-report": {
      "post": {
        "summary": "Generate Report",
        "description": "Generate attendance report and return JSON with download URL\n\nParameters:\n- company_code: Company code for login (default: 85)\n- username: Username for login (default: hisham.octacer)\n- password: Password for login\n- report_date: Date for report in YYYY-MM-DD format (default: today)\n\nReturns:\n- JSON with report_url (download link), file_id, and metadata",
        "operationId": "generate_report_generate_report_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReportRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReportResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/generate-report-direct": {
      "post": {
        "summary": "Generate Report Direct",
        "description": "Generate attendance report and directly return Excel file\n\nParameters:\n- company_code: Company code for login (default: 85)\n- username: Username for login (default: hisham.octacer)\n- password: Password for login\n- report_date: Date for report in YYYY-MM-DD format (default: today)\n\nReturns:\n- Excel file with attendance report (direct download)",
        "operationId": "generate_report_direct_generate_report_direct_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReportRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/download/{file_id}": {
      "get": {
        "summary": "Download File",
        "description": "Download a generated report by file ID\n\nParameters:\n- file_id: The unique file ID returned from generate-report endpoint\n\nReturns:\n- Excel file download",
        "operationId": "download_file_download__file_id__get",
        "parameters": [
          {
            "name": "file_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "File Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/debug/{debug_id}": {
      "get": {
        "summary": "Get Debug Info",
        "description": "Get debug information and list of debug files for a specific debug ID\n\nParameters:\n- debug_id: The debug ID returned in error response\n\nReturns:\n- JSON with list of available debug files and download URLs",
        "operationId": "get_debug_info_debug__debug_id__get",
        "parameters": [
          {
            "name": "debug_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Debug Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ReportRequest": {
        "properties": {
          "company_code": {
            "type": "string",
            "title": "Company Code",
            "default": "85"
          },
          "username": {
            "type": "string",
            "title": "Username",
            "default": "hisham.octacer"
          },
          "password": {
            "type": "string",
            "title": "Password",
            "default": "P@ss1234"
          },
          "report_date": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Report Date"
          }
        },
        "type": "object",
        "title": "ReportRequest"
      },
      "ReportResponse": {
        "properties": {
          "report_url": {
            "type": "string",
            "title": "Report Url"
          },
          "file_id": {
            "type": "string",
            "title": "File Id"
          },
          "report_date": {
            "type": "string",
            "title": "Report Date"
          },
          "generated_at": {
            "type": "string",
            "title": "Generated At"
          },
          "expires_in": {
            "type": "integer",
            "title": "Expires In"
          }
        },
        "type": "object",
        "required": [
          "report_url",
          "file_id",
          "report_date",
          "generated_at",
          "expires_in"
        ],
        "title": "ReportResponse"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}